AC - tests: extract shared MCP harness to tests/common/mod.rs (fixes #38)
The MCP test harness — bin_path() + Harness struct + connect() +
call() — was duplicated across 7 integration test files (character_core,
checks, combat, inventory, npcs, setup, world). ~60 lines × 7 = 420
duplicated lines. Drift was already showing — inventory.rs had
_db_path (unused) where the others had db_path (used).

Cargo treats tests/common/mod.rs as a non-test module shared across
integration test binaries. Each test file now adds two lines at the
top:

    mod common;
    use common::{call, connect};

Net change: -284 lines. tests/common/mod.rs is 102 lines (well-doc'd);
the 7 test files each lose ~60 lines of boilerplate.

Per-file specifics that vary (make_char with different signatures,
setup_world bootstrap, kill_character helper, latest_event_payload
DB inspector) stay in their respective files — extraction would have
required either parametrising every helper or accepting the variation
as harness divergence.

Files NOT extracted:
- tests/healthz.rs uses TCP / reqwest, not the rmcp client harness
- tests/dice.rs has its own simpler call_roll wrapper
- tests/content_introspect.rs has a connect_with_db variant
- tests/schema.rs spawns the binary as an HTTP child for schema
  inspection
- tests/stdio_server_info.rs has minimal helpers

These could each get tests/common helpers in follow-ups, but their
patterns vary enough that the simple bin_path/Harness/connect/call
shape doesn't fit cleanly.

Behavior unchanged — all 13 test binaries pass without modification.
The Harness struct's _tmp/db_path/client fields are now pub so test
files can access h.db_path for DB inspection (matches existing usage).
